feat: replace everything with new dispatcher
