--- crates/bitwarden-threading/src/call_target.rs
@@ -1,8 +1,0 @@
-use std::future::Future;
-
-pub trait CallTarget {
-    type Input: Send + 'static;
-    type Output: Send + 'static;
-
-    fn call(&self, input: Self::Input) -> impl Future<Output = Self::Output>;
-}
