--- crates/bitwarden-threading/src/traits/mod.rs
@@ -1,1 +1,0 @@
-
