--- crates/bitwarden-threading/src/wasm/mod.rs
@@ -1,1 +1,0 @@
-
